Добавил проверку на наличие request аргумента в save #58

+++ backend/plugins/addToHistory.js
@@ -5,15 +5,17 @@ module.exports = function (schema) {
     let user, type;
 
     schema.pre('save', function (next, request) {
-        user = request.currentUser;
+        if(typeof request === 'object') {
+            user = request.currentUser;
+            if(request.isRemoved) type = 'delete';
+        }
 
         type = this.isNew ? 'add' : 'edit';
-        if(request.isRemoved) type = 'delete';
 
         next()
     })
 
-    schema.pre('save', async function (info) {
+    schema.post('save', async function (info) {
         await History.create({ user, info, type })
     })
